@@ -236,8 +236,8 @@ class Dataset(torch.utils.data.Dataset):
 
     def _process(self):
         f = osp.join(self.processed_dir, 'pre_transform.pt')
-        if osp.exists(f) and torch.load(f, weights_only=False) != _repr(
-                self.pre_transform):
+        if not self.force_reload and osp.exists(f) and torch.load(
+                f, weights_only=False) != _repr(self.pre_transform):
             warnings.warn(
                 "The `pre_transform` argument differs from the one used in "
                 "the pre-processed version of this dataset. If you want to "
@@ -246,8 +246,8 @@ class Dataset(torch.utils.data.Dataset):
                 stacklevel=2)
 
         f = osp.join(self.processed_dir, 'pre_filter.pt')
-        if osp.exists(f) and torch.load(f, weights_only=False) != _repr(
-                self.pre_filter):
+        if not self.force_reload and osp.exists(f) and torch.load(
+                f, weights_only=False) != _repr(self.pre_filter):
             warnings.warn(
                 "The `pre_filter` argument differs from the one used in "
                 "the pre-processed version of this dataset. If you want to "
