@@ -79,7 +79,7 @@ class ModelNet(InMemoryDataset):
 
     urls = {
         '10':
-        'http://vision.princeton.edu/projects/2014/3DShapeNets/ModelNet10.zip',
+        'http://3dvision.princeton.edu/projects/2014/3DShapeNets/ModelNet10.zip',  # noqa
         '40': 'http://modelnet.cs.princeton.edu/ModelNet40.zip'
     }
 
