@@ -8,6 +8,7 @@ from .encoding import PositionalEncoding, TemporalEncoding
 from .summary import summary
 
 from .aggr import *  # noqa
+from .attention import *  # noqa
 from .conv import *  # noqa
 from .pool import *  # noqa
 from .glob import *  # noqa
@@ -17,7 +18,6 @@ from .dense import *  # noqa
 from .kge import *  # noqa
 from .models import *  # noqa
 from .functional import *  # noqa
-from .nlp import *  # noqa
 
 __all__ = [
     'Reshape',
