@@ -3,7 +3,7 @@ from .qformer import QFormer
 from .sgformer import SGFormerAttention
 from .polynormer import PolynormerAttention
 
-__all__ = [
+__all__ = classes = [
     'PerformerAttention',
     'QFormer',
     'SGFormerAttention',
