@@ -29,12 +29,9 @@ from .gnnff import GNNFF
 from .pmlp import PMLP
 from .neural_fingerprint import NeuralFingerprint
 from .visnet import ViSNet
-from .g_retriever import GRetriever
-from .git_mol import GITMol
-from .molecule_gpt import MoleculeGPT
-from .protein_mpnn import ProteinMPNN
-from .glem import GLEM
+from .lpformer import LPFormer
 from .sgformer import SGFormer
+
 from .polynormer import Polynormer
 # Deprecated:
 from torch_geometric.explain.algorithm.captum import (to_captum_input,
@@ -85,11 +82,7 @@ __all__ = classes = [
     'PMLP',
     'NeuralFingerprint',
     'ViSNet',
-    'GRetriever',
-    'GITMol',
-    'MoleculeGPT',
-    'ProteinMPNN',
-    'GLEM',
+    'LPFormer',
     'SGFormer',
     'Polynormer',
     'ARLinkPredictor',
